basic refactor existing code to use spdx go tools for decoding
