+++ internal/formats/spdx22tagvalue/decoder.go
@@ -1,17 +1,17 @@
-package spdx22json
+package spdx22tagvalue
 
 import (
 	"fmt"
 	"io"
 
-	"github.com/spdx/tools-golang/jsonloader"
+	"github.com/spdx/tools-golang/tvloader"
 
 	"github.com/anchore/syft/internal/formats/common/spdxhelpers"
 	"github.com/anchore/syft/syft/sbom"
 )
 
 func decoder(reader io.Reader) (*sbom.SBOM, error) {
-	doc, err := jsonloader.Load2_2(reader)
+	doc, err := tvloader.Load2_2(reader)
 	if err != nil {
 		return nil, fmt.Errorf("unable to decode spdx-json: %w", err)
 	}
