+++ internal/formats/spdx22tagvalue/validator.go
@@ -1,4 +1,4 @@
-package spdx22json
+package spdx22tagvalue
 
 import (
 	"io"
